Add files via upload
This is one way we could clean the data, it takes the averages of the two teams playing (calculated from previous games in the season+last 3 games of the previous season), takes into consideration information we could know about the game before the game starts, and matches that data to the winner of the game. In said data the first group of data is home team averages, the second grouping is away team averages, and the last grouping is game stats, with the y being 1 for hometeam victory and 0 for awayteam victory. The models usually give >0.5 for a score.